test: reuse command helper across CLI and release checks
